requirements.txt
discord
python-dotenv
requests
lxml
apscheduler